move to component factory resolver and add Toast to ngmodule
