+++ src/components/toastr/toastr.ts
@@ -8,12 +8,7 @@ export {
 export { Overlay } from './overlay/overlay';
 export { OverlayContainer } from './overlay/overlay-container';
 export { OverlayRef } from './overlay/overlay-ref';
-
 export { Toast } from './toast-component';
-export {
-  ToastrService,
-  TOASTR_PROVIDERS,
-  ActiveToast,
-  ToastrModule
-} from './toastr-service';
+export { ToastrService, ActiveToast } from './toastr-service';
 export { ToastrConfig, ToastConfig } from './toastr-config';
+export { ToastrModule, TOASTR_PROVIDERS } from './toastr-module';
